Refactoring event handling
Instead of using the somewhat magical "event_multiplexer" mechanism, use
plain and simple message passing instead.

Remove the need for `EventHandler` `DispatchTo` trait bounds.

Now states needed for handling these messages are mostly owned by the
event handling tasks. So there is also no need for the `State` trait
anymore.

To support the event handling tasks, now the object store, the client
connection and the server context are cloneable and shareable across
tasks. And to keep the object store consistent across awaits, the store
is now lockable, preventing other tasks from changing the store when
needed. This is needed as there are tasks running concurrently now.

Also added `Objects::by_interface` to iterate over objects of a certain
interface in the object stores. To remove the need of tracking objects
of a certain type from the event handling tasks.

There are also some trait definition tweaks, like splitting
`ObjectMeta` out of `Object<Ctx>`, and `GlobalMeta` out of `Bind<Ctx>`,
so the `Ctx` parameter doesn't pollute as much.

+++ wl_io/src/lib.rs
@@ -184,7 +184,7 @@ impl Iterator for FdIter<'_> {
     fn next(&mut self) -> Option<Self::Item> {
         let p = unsafe { libc::CMSG_DATA(self.cmsghdr as *const _) };
         let data_len =
-            self.cmsghdr as *const _ as usize + self.cmsghdr.cmsg_len as usize - p as usize;
+            self.cmsghdr as *const _ as usize + self.cmsghdr.cmsg_len - p as usize;
         let nfds = data_len / std::mem::size_of::<RawFd>();
         let fds = unsafe { std::slice::from_raw_parts(p as *const RawFd, nfds) };
         let ret = fds.get(self.idx).copied();
@@ -209,9 +209,9 @@ where
             cmsg_buffer
                 .as_mut()
                 .unwrap()
-                .set_len(mhdr.msg_controllen as usize);
+                .set_len(mhdr.msg_controllen);
             debug_assert!(!mhdr.msg_control.is_null());
-            debug_assert!(msg_controllen >= mhdr.msg_controllen as usize);
+            debug_assert!(msg_controllen >= mhdr.msg_controllen);
             libc::CMSG_FIRSTHDR(&mhdr as *const libc::msghdr)
         } else {
             std::ptr::null()
